Add support for Linux (acpi) and restructure the project
- Use plugins to implement power monitors
- Currently supported plugins:
  - macOS:
    - `pmset` (polling)
    - `IOPowerSources` C API (event loop, more efficient than polling)
  - Linux:
    - `acpi` (polling)
- A template for `upower` (Linux, event-based) is added but still needs
  to be implemented

+++ power_saver.tmux
@@ -4,14 +4,20 @@ CURRENT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 SCRIPTS_DIR="${CURRENT_DIR}/scripts"
 
 
-main() {
+kill_previous_instances () {
     local previous_processes pid
 
-    # Ensure previous instances are killed before a new is started
     mapfile -t previous_processes <<< "$(ps -e -o "pid=,command=" | grep 'monitor_power_supply\.sh' | awk '{ print $1 }')"
+    [[ "${#previous_processes}" -gt 0 ]] || return 0
     for pid in "${previous_processes[@]}"; do
         kill "${pid}"
     done
+}
+
+
+main() {
+    # Ensure previous instances are killed before a new is started
+    kill_previous_instances
     "${SCRIPTS_DIR}/monitor_power_supply.sh" &
 }
